Extracts detect manifests functionality into separate class
